Balance management
1.  I add the column “balance” which is an integer type.
2. For the deposit function, I use the charge.create method to check
the stripe token. (I pass the stripe_token to the attribute “source”)
3.  For the transfer function, I use just reduce the source user’s
amount and increase the target user’s amount
4. For the withdraw function, I use the recipient.create() method, and
pass the stripe_token as a source.
5. Still for the withdraw function, I find that there is no check for
the stripe_token in the api, but I have checked in the code, you can
decide whether to have it or not.

Ps: I haven’t test this controller, maybe I will do that during the
vacation. I just  publish this code now if you need to use it. There
may be bugs in this branch, if you don’t mind I will fix them later.